Move hud_menu_static into menus folder
