+++ mp/src/game/client/hl2/hud_weaponselection.cpp
@@ -7,7 +7,7 @@
 #include "cbase.h"
 #include "weapon_selection.h"
 #include "iclientmode.h"
-#include "../momentum/ui/HUD/hud_menu_static.h"
+#include "../momentum/ui/menus/hud_menu_static.h"
 #include "history_resource.h"
 #include "input.h"
 #include "../hud_crosshair.h"
